% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccmSummary.R
\name{adj}
\alias{adj}
\title{Adjacency Matrix}
\usage{
adj(mat, thresh = 0.001)
}
\arguments{
\item{mat}{Numeric matrix.}

\item{thresh}{Threshold for absolute value of entries.}
}
\value{
An adjacency matrix containing 1's and 0's.
}
\description{
This function calculates an adjacency matrix for the
matrix \eqn{mat} based on the absolute value threshold of \eqn{thresh}.
}
