% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccmSummary.R
\name{aicbicm}
\alias{aicbicm}
\title{Modified AIC and BIC}
\usage{
aicbicm(x, omegaks, omega0s, lambda2, ws)
}
\arguments{
\item{x}{List of \eqn{K} data matrices each of dimension \eqn{n_k} x \eqn{p}.}

\item{omegaks}{\eqn{p} x \eqn{p} x \eqn{K} array of \eqn{K} number of estimated subject-level precision matrices.}

\item{omega0s}{\eqn{p} x \eqn{p} x \eqn{nclusts} array of \eqn{nclusts} number of estimated cluster-level precision matrices.}

\item{lambda2}{Scalar tuning parameter value used for RCCM.}

\item{ws}{\eqn{nclusts} x \eqn{K} matrix of estimated cluster weights for each subject.}
}
\value{
Numeric vector of length 2 containing the modified AIC and BIC values.
}
\description{
This function calculates modified AIC and BIC values
for the random covariance clustering model (RCCM)
}
\examples{
# Generate data
set.seed(1994)
myData <- rccSim(G = 2, clustSize = 10, p = 10, n = 100, overlap = 0.50, rho = 0.10)

# Analyze with RCCM
resultRccm <- rccm(x = myData$simDat, lambda1 = 20,
lambda2 = 325, lambda3 = 0.01, nclusts = 2)

# Calculate modified AIC and BIC
aicbicm(x = myData$simDat, omegaks = resultRccm$Omegas,
omega0s = resultRccm$Omega0, lambda2 = 325, ws = resultRccm$weights)

}
