% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccmSummary.R
\name{bic_cal}
\alias{bic_cal}
\title{BIC for RCM}
\usage{
bic_cal(x, Omegas, Gk_est = NULL)
}
\arguments{
\item{x}{List of \eqn{K} data matrices each of dimension \eqn{n_k} x \eqn{p}.}

\item{Omegas}{\eqn{p} x \eqn{p} x \eqn{K} array of \eqn{K} number of estimated subject-level precision matrices.}

\item{Gk_est}{\eqn{p} x \eqn{p} x \eqn{K} array of \eqn{K} number of estimated subject-level networks.}
}
\value{
Numeric BIC value
}
\description{
This function calculates the BIC for the random covariance model (RCM)
}
\examples{
# Generate data
set.seed(1994)
myData <- rccSim(G = 1, clustSize = 10, p = 10, n = 100, overlap = 0.50, rho = 0.10)

# Analyze with RCM
rcmRes <- randCov(myData$simDat, lambda1 = 0.01, lambda2 = 0.01, lambda3 = 0, delta = 0.0001)

# Calculate BIC for the RCM
bic_cal(x = myData$simDat, Omegas = rcmRes$Omegas)

}
\author{
Lin Zhang
}
