% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccmSummary.R
\name{dwishart}
\alias{dwishart}
\title{Probability Density Function for Wishart Distribution}
\usage{
dwishart(x, M, nu, logged = FALSE)
}
\arguments{
\item{x}{\eqn{p} x \eqn{p} positive definite matrix.}

\item{M}{\eqn{p} x \eqn{p} mean matrix. Note that \eqn{M = nu*V} where \eqn{V} is the scale matrix.}

\item{nu}{Degrees of freedom.}

\item{logged}{Logical. If TRUE, probability given on log scale.}
}
\value{
The probability density function value.
}
\description{
This function provides the probability density function for
the Wishart distribution.
}
