% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvTune.R
\name{MatClust}
\alias{MatClust}
\title{Matrix Clustering}
\usage{
MatClust(mats, G)
}
\arguments{
\item{mats}{List of \eqn{K} data matrices each of dimension \eqn{n_k} x \eqn{p}.}

\item{G}{Number of groups or clusters.}
}
\value{
\eqn{G} x \eqn{K} matrix of weights
}
\description{
This function provides the probability density function for
the Wishart distribution.
}
