% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccmSummary.R
\name{mbic_cal}
\alias{mbic_cal}
\title{Modified BIC for RCM}
\usage{
mbic_cal(x, Omega0, Omegas, lambda2, G0_est = NULL, Gk_est = NULL)
}
\arguments{
\item{x}{List of \eqn{K} data matrices each of dimension \eqn{n_k} x \eqn{p}.}

\item{Omega0}{\eqn{p} x \eqn{p} group-level precision matrix estimate.}

\item{Omegas}{\eqn{p} x \eqn{p} x \eqn{K} array of \eqn{K} number of estimated subject-level precision matrices.}

\item{lambda2}{Non-negative scalar. Induces similarity between subject-level matrices and group-level matrix.}

\item{G0_est}{\eqn{p} x \eqn{p} group-level network estimate.}

\item{Gk_est}{\eqn{p} x \eqn{p} x \eqn{K} array of \eqn{K} number of estimated subject-level networks.}
}
\value{
Numeric Modified BIC value
}
\description{
This function calculates the modified BIC for the random covariance model (RCM)
}
\examples{
# Generate data
set.seed(1994)
myData <- rccSim(G = 1, clustSize = 10, p = 10, n = 100, overlap = 0.50, rho = 0.10)

# Analyze with RCM
rcmRes <- randCov(myData$simDat, lambda1 = 0.01, lambda2 = 0.01, lambda3 = 0, delta = 0.0001)

# Calculate modified BIC for the RCM
mbic_cal(x = myData$simDat, Omega0 = rcmRes$Omega0, Omegas = rcmRes$Omegas,
         lambda2 = 0.01)

}
\author{
Lin Zhang
}
