% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvTune.R
\name{nllikCalc}
\alias{nllikCalc}
\title{Negative Log-Likelihood}
\usage{
nllikCalc(omegaks, datf)
}
\arguments{
\item{omegaks}{\eqn{p} x \eqn{p} x \eqn{K} array of \eqn{K} number of estimated subject-level precision matrices.}

\item{datf}{List of \eqn{K} data matrices each of dimension \eqn{n_k} x \eqn{p}.}
}
\value{
Negative log-likelihood
}
\description{
This function calculates the negative log-likelihood for
selecting the optimal tuning parameters using cross-validation.
}
