% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccmSummary.R
\name{randCalc}
\alias{randCalc}
\title{Rand Index}
\usage{
randCalc(x, y)
}
\arguments{
\item{x}{First integer vector of cluster memberships.}

\item{y}{Second integer vector of cluster memberships.}
}
\value{
The rand index value, bounded between 0 and 1.
}
\description{
This function calculates the rand index describing
the amount of agreement between two integer vectors
 of cluster memberships.
}
