% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randCov.R
\name{randCov}
\alias{randCov}
\title{Random Covariance Model}
\usage{
randCov(x, lambda1, lambda2, lambda3 = 0, delta = 0.001,
  max.iters = 100)
}
\arguments{
\item{x}{List of \eqn{K} data matrices each of dimension \eqn{n_k} x \eqn{p}.}

\item{lambda1}{Non-negative scalar. Induces sparsity in subject-level matrices.}

\item{lambda2}{Non-negative scalar. Induces similarity between subject-level matrices and group-level matrix.}

\item{lambda3}{Non-negative scalar. Induces sparsity in group-level matrix.}

\item{delta}{Threshold for convergence.}

\item{max.iters}{Maximum number of iterations for block coordinate descent optimization.}
}
\value{
A list of length 2 containing:
\enumerate{
\item Group-level precision matrix estimate (Omega0).
\item \eqn{p} x \eqn{p} x \eqn{K} array of \eqn{K} subject-level precision matrix estimates (Omegas).
}
}
\description{
This function implements the Random Covariance Model (RCM) for joint estimation of
multiple sparse precision matrices. Optimization is conducted using block
coordinate descent.
}
\examples{
# Generate data with 5 subjects, 15 variables for each subject,
# 100 observations for each variable for each subject,
# and 10\% of differential connections
# within each group
myData <- rccSim(G = 1, clustSize = 5, p = 15, n = 100, rho = 0.10)

# Analyze simulated data with RCM
result <- randCov(x = myData$simDat, lambda1 = 0.30, lambda2 = 0.10, lambda3 = 0.001, delta = 0.001)

}
\references{
Zhang, Lin, Andrew DiLernia, Karina Quevedo, Jazmin Camchong, Kelvin Lim, and Wei Pan.
"A Random Covariance Model for Bi-level Graphical Modeling with Application to Resting-state FMRI Data." 2019. https://arxiv.org/pdf/1910.00103.pdf
}
\author{
Lin Zhang and Andrew DiLernia
}
