% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccmSummary.R
\name{rccmLogLike}
\alias{rccmLogLike}
\title{Model Log-Likelihood}
\usage{
rccmLogLike(omegaks, omega0s, ws, x, lambda2)
}
\arguments{
\item{omegaks}{\eqn{p} x \eqn{p} x \eqn{K} array of \eqn{K} number of estimated subject-level precision matrices.}

\item{omega0s}{\eqn{p} x \eqn{p} x \eqn{nclusts} array of \eqn{nclusts} number of estimated cluster-level precision matrices.}

\item{ws}{\eqn{nclusts} x \eqn{K} matrix of estimated cluster weights for each subject (weights).}

\item{x}{List of \eqn{K} data matrices each of dimension \eqn{n_k} x \eqn{p}.}

\item{lambda2}{Non-negative scalar value used as input to rccm function to obtain estimates.}
}
\value{
Numeric AIC value.
}
\description{
This function calculates the model log-likelihood
for the random covariance clustering model (RCCM)
}
\examples{
# Generate data
set.seed(1994)
myData <- rccSim(G = 2, clustSize = 10, p = 10, n = 100, overlap = 0.50, rho = 0.10)

# Analyze with RCCM
resultRccm <- rccm(x = myData$simDat, lambda1 = 20,
lambda2 = 325, lambda3 = 0.01, nclusts = 2)

# Calculate AIC
rccmLogLike(omegaks = resultRccm$Omegas, omega0s = resultRccm$Omega0,
ws = resultRccm$weights, x = myData$simDat, lambda2 = 325)

}
