% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccSim.R
\name{rccSim}
\alias{rccSim}
\title{Simulation of Data for RCCM}
\usage{
rccSim(
  G = 2,
  clustSize = c(67, 37),
  p = 10,
  n = 177,
  overlap = 0.5,
  rho = 0.1,
  esd = 0.05,
  type = "hub",
  eprob = 0.5
)
}
\arguments{
\item{G}{Positive integer. Number of groups or clusters.}

\item{clustSize}{Positive integer or vector of positive integers. Number of subjects in each cluster.}

\item{p}{Positive integer. Number of variables for each subject.}

\item{n}{Positive integer. Number of observations for each subject on each variable.}

\item{overlap}{Positive number between 0 and 1. Approximate proportion of overlapping edges across cluster-level networks.}

\item{rho}{Positive number between 0 and 1. Approximate proportion of differential edges for subjects compared to their corresponding cluster-level network.}

\item{esd}{Standard deviation of mean 0 noise added to generated subject-level matrices for variation from the corresponding group-level matrix.}

\item{type}{Graph type. Options are "hub" or "random".}

\item{eprob}{Probability of two nodes having an edge between them. Only applicable if type = "random".}
}
\value{
A list of length 5 containing:
\enumerate{
\item list of \eqn{K} multivariate-Gaussian data sets each of dimension \eqn{n_k} x \eqn{p} (simDat).
\item \eqn{p} x \eqn{p} x \eqn{G} array of \eqn{G} number of cluster-level networks (g0s).
\item \eqn{p} x \eqn{p} x \eqn{G} array of \eqn{G} number of cluster-level precision matrices (Omega0s).
\item \eqn{p} x \eqn{p} x \eqn{K} array of \eqn{K} number of subject-level precision matrices (Omegaks).
\item vector of length \eqn{K} containing cluster memberships for each subject (zgks).
}
}
\description{
This function simulates data based on the Random Covariance Clustering Model (RCCM).
Data is generated in a hierarchical manner, beginning with group-level networks
and precision matrices and then subject-level networks and matrices.
}
\details{
For simulating data for hub type graphs, \eqn{G} cluster-level networks are first generated,
each with floor(\eqn{\sqrt p})
hubs and thus \eqn{E = p -} floor(\eqn{\sqrt p}) edges.  For generating random graphs,
\eqn{G} cluster-level networks are generating such that nodes are connected
with a probability specified by \eqn{eprob}, yielding approximately \eqn{E=} (\eqn{p} choose 2) \eqn{x eprob} edges.
Cluster-level networks are forced
to share \eqn{s =} floor(\eqn{overlap} x \eqn{E}) edges.
Note that \eqn{overlap} represents the approximate proportion
of edges that are common across the cluster-level networks.

Then, for the \eqn{K} subject-level matrices, we first randomly
assign them to the \eqn{G} clusters, and then subject-level
networks are generated by randomly selecting
floor(\eqn{rho} x \eqn{E}) node pairs to add or remove an edge
from their corresponding cluster-level network.
Non-zero entries for all precision matrices are generated from
a uniform distribution with support on the interval
[-1, -0.50] U [0.50, 1],
and are adjusted until positive definite matrices are obtained.
}
\examples{
# Generate data with 2 clusters with 12 and 10 subjects respectively,
# 15 variables for each subject, 100 observations for each variable for each subject,
# the groups sharing about 50\% of network connections, and 10\% of differential connections
# within each group
myData <- rccSim(G = 2, clustSize = c(12, 10), p = 15, n = 100, overlap = 0.50, rho = 0.10)

# View list of simulated data
View(myData)

}
\author{
Andrew DiLernia
}
