% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starsRccm.R
\name{starsRccm}
\alias{starsRccm}
\title{Modified Stability Approach for Regularization Selection}
\usage{
starsRccm(
  datf,
  lambs,
  method = "RCCM",
  G = 2,
  N = 10,
  beta = 0.05,
  z0s = NULL,
  ncores = NULL
)
}
\arguments{
\item{datf}{List of \eqn{K} data sets each of dimension \eqn{n_k} x \eqn{p}.}

\item{lambs}{A data frame of candidate tuning parameter values with three columns: lambda1, lambda2, and lambda3.}

\item{method}{Method to implement modified stARS algorithm for. Must be one of "FGL", "GGL", "GLasso", "RCCM", or "RCM".}

\item{G}{Number of groups or clusters. Only applicable if method = "RCCM".}

\item{N}{Number of subsamples for modified stARS algorithm}

\item{beta}{Positive scalar between 0 and 1. Limits allowed amount of instability across subsamples.}

\item{z0s}{Vector of length \eqn{K} with initial cluster memberships. Only applicable if method = "RCCM".}

\item{ncores}{Number of computing cores to use if desired to run in parallel. Optional.}
}
\value{
A data frame of optimally selected tuning parameter values and the sparsity level with four columns: lambda1, lambda2, lambda3, and sparsity.
}
\description{
This function implements a modified stability approach for
regularization selection (stARS) method for tuning parameter
selection. Methods available to implement include the
fused graphical lasso (\link[JGL:JGL]{FGL}), group graphical lasso (\link[JGL:JGL]{GGL}),
graphical lasso (\link[glasso:glasso]{GLasso}), random covariance clustering
model (\link[rcm:rccm]{RCCM}), and the random covariance model (\link[rcm:randCov]{RCM}).
}
\examples{
# Generate data with 2 clusters with 10 subjects in each group,
# 10 variables for each subject, 100 observations for each variable for each subject,
# the groups sharing about 50\% of network connections, and 10\% of differential connections
# within each group
set.seed(1994)
myData <- rccSim(G = 2, clustSize = 10, p = 10, n = 100, overlap = 0.50, rho = 0.10)

# Find optimal tuning parameter set using modified stARS
optTune <- starsRccm(datf = myData$simDat, lambs = expand.grid(lambda1 = c(20, 25, 30),
lambda2 = c(300, 325), lambda3 = 0.01), method = "RCCM", G = 2)

# Analyze with RCCM using optimally selected tuning parameters
resultRccm <- rccm(x = myData$simDat, lambda1 = optTune$lambda1[1],
lambda2 = optTune$lambda2[1], lambda3 = optTune$lambda3[1], nclusts = 2)

}
\references{
Liu, Han, Kathryn Roeder, and Larry Wasserman.
"Stability Approach to Regularization Selection (StARS) for High Dimensional Graphical Models." 2010.

Danaher, Patrick, Pei Wang, and Daniela M. Witten.
"The Joint Graphical Lasso for Inverse Covariance Estimation across Multiple Classes." Journal of the Royal Statistical Society: Series B (Statistical Methodology) 76, no. 2 (2014): 373-97.

Friedman, Jerome, Trevor Hastie, and Robert Tibshirani.
"Sparse Inverse Covariance Estimation with the Graphical Lasso." Biostatistics 9, no. 3 (2008): 432-41.

Zhang, Lin, Andrew DiLernia, Karina Quevedo, Jazmin Camchong, Kelvin Lim, and Wei Pan.
"A Random Covariance Model for Bi-level Graphical Modeling with Application to Resting-state FMRI Data." 2019. https://arxiv.org/pdf/1910.00103.pdf
}
\author{
Andrew DiLernia
}
