% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccmSummary.R
\name{zToA}
\alias{zToA}
\title{Z to Adjacency Matrix}
\usage{
zToA(z)
}
\arguments{
\item{z}{Integer vector of cluster memberships.}
}
\value{
An adjacency matrix containing 1's and 0's.
}
\description{
This function calculates an adjacency matrix based on
an integer vector of cluster memberships.
}
\examples{
# Calculate adjacency matrix for clustering
zToA(c(rep(1, 2), rep(2, 2), rep(3, 2)))

}
